Bug 1577236 - clang-10: Fix -Wimplicit-int-float-conversion warnings in dom/animation/ r=hiro
Depends on D43779

Differential Revision: https://phabricator.services.mozilla.com/D43780

UltraBlame original commit: 9dcc8e550559887b04d810e828eebc6a2ab2027d

+++ dom/animation/AnimationEffect.cpp
@@ -182,9 +182,9 @@ ComputedTiming AnimationEffect::GetComputedTimingAt(
   
   
   result.mCurrentIteration =
-      (result.mIterations >= UINT64_MAX &&
+      (result.mIterations >= double(UINT64_MAX) &&
        result.mPhase == ComputedTiming::AnimationPhase::After) ||
-              overallProgress >= UINT64_MAX
+              overallProgress >= double(UINT64_MAX)
           ? UINT64_MAX  
                         
           : static_cast<uint64_t>(overallProgress);
